+++ core/src/test/java/com/alibaba/nacos/core/remote/grpc/GrpcServerTest.java
@@ -18,9 +18,10 @@
 package com.alibaba.nacos.core.remote.grpc;
 
 import com.alibaba.nacos.common.remote.ConnectionType;
-import com.alibaba.nacos.core.remote.RpcServerTlsConfig;
+import com.alibaba.nacos.core.remote.tls.RpcServerTlsConfig;
 import com.alibaba.nacos.sys.env.EnvUtil;
 import com.alibaba.nacos.sys.utils.ApplicationUtils;
+import io.grpc.netty.shaded.io.grpc.netty.InternalProtocolNegotiator;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -88,10 +89,20 @@ public class GrpcServerTest {
     public void testGrpcEnableTls() throws Exception {
         final BaseGrpcServer grpcSdkServer = new BaseGrpcServer() {
             @Override
+            protected InternalProtocolNegotiator.ProtocolNegotiator newProtocolNegotiator() {
+                return null;
+            }
+    
+            @Override
             public ThreadPoolExecutor getRpcExecutor() {
                 return null;
             }
-            
+    
+            @Override
+            public void reloadSslContext() {
+        
+            }
+    
             @Override
             public int rpcPortOffset() {
                 return 100;
@@ -114,10 +125,20 @@ public class GrpcServerTest {
         
         final BaseGrpcServer grpcSdkServer = new BaseGrpcServer() {
             @Override
+            protected InternalProtocolNegotiator.ProtocolNegotiator newProtocolNegotiator() {
+                return null;
+            }
+    
+            @Override
             public ThreadPoolExecutor getRpcExecutor() {
                 return null;
             }
-            
+    
+            @Override
+            public void reloadSslContext() {
+        
+            }
+    
             @Override
             public int rpcPortOffset() {
                 return 100;
@@ -139,10 +160,20 @@ public class GrpcServerTest {
     public void testGrpcEnableMutualAuthAndPart() throws Exception {
         final BaseGrpcServer grpcSdkServer = new BaseGrpcServer() {
             @Override
+            protected InternalProtocolNegotiator.ProtocolNegotiator newProtocolNegotiator() {
+                return null;
+            }
+    
+            @Override
             public ThreadPoolExecutor getRpcExecutor() {
                 return null;
             }
-            
+    
+            @Override
+            public void reloadSslContext() {
+        
+            }
+    
             @Override
             public int rpcPortOffset() {
                 return 100;
